+++ eZ/Publish/Core/REST/Common/FieldTypeProcessor/RelationProcessor.php
@@ -10,10 +10,6 @@
  */
 namespace eZ\Publish\Core\REST\Common\FieldTypeProcessor;
 
-use eZ\Publish\Core\REST\Common\FieldTypeProcessor;
-use eZ\Publish\Core\FieldType\Relation\Type;
-use Symfony\Component\Routing\RouterInterface;
-
 class RelationProcessor extends BaseRelationProcessor
 {
     /**
@@ -34,41 +30,5 @@ class RelationProcessor extends BaseRelationProcessor
         );
 
         return $outgoingValueHash;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function preProcessFieldSettingsHash($incomingSettingsHash)
-    {
-        if (isset($incomingSettingsHash['selectionMethod'])) {
-            switch ($incomingSettingsHash['selectionMethod']) {
-                case 'SELECTION_BROWSE':
-                    $incomingSettingsHash['selectionMethod'] = Type::SELECTION_BROWSE;
-                    break;
-                case 'SELECTION_DROPDOWN':
-                    $incomingSettingsHash['selectionMethod'] = Type::SELECTION_DROPDOWN;
-            }
-        }
-
-        return $incomingSettingsHash;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function postProcessFieldSettingsHash($outgoingSettingsHash)
-    {
-        if (isset($outgoingSettingsHash['selectionMethod'])) {
-            switch ($outgoingSettingsHash['selectionMethod']) {
-                case Type::SELECTION_BROWSE:
-                    $outgoingSettingsHash['selectionMethod'] = 'SELECTION_BROWSE';
-                    break;
-                case Type::SELECTION_DROPDOWN:
-                    $outgoingSettingsHash['selectionMethod'] = 'SELECTION_DROPDOWN';
-            }
-        }
-
-        return $outgoingSettingsHash;
     }
 }
